refractored permission registering to register on message create event and interaction create event
