+++ src/botfiles/events/ready/ReadyEvent.ts
@@ -3,6 +3,7 @@ import DiscordClient from '../../client/client';
 import { SlashCommandBuilder } from '@discordjs/builders';
 import { PermissionResolvable, ApplicationCommandPermissionData } from 'discord.js';
 import logger from '../../../config/logger';
+import { registerPermissions } from '../../utils/registry';
 const { REST } = require('@discordjs/rest');
 const { Routes } = require('discord-api-types/v9');
 
@@ -24,42 +25,10 @@ export default class ReadyEvent extends BaseEvent {
         try {
             logger.info('Bot has logged in.');
             const guilds = await client.guilds.fetch();
-            const permissions: RolePermission[] = [];
-            const fullPermissions: any[] = [];
+            // const permissions: RolePermission[] = [];
+            // const fullPermissions: any[] = [];
             guilds.forEach(async (guild) => {
-                const tempGuild = await guild.fetch();
-                client.commands.forEach(async (command) => {
-                    if (command.getPermissions().length === 0) return;
-                    const roles = tempGuild.roles.cache.filter((r) =>
-                        r.permissions.has(command.getPermissions() as PermissionResolvable)
-                    );
-
-                    roles.forEach((r) => {
-                        permissions.push({
-                            id: r.id,
-                            type: 'ROLE',
-                            permission: true
-                        });
-                    });
-                });
-                const commands = await client.application?.commands.fetch({
-                    guildId: guild.id
-                });
-
-                commands?.forEach(async (cmd) => {
-                    const tempCommand = client.commands.get(cmd.name);
-
-                    if (tempCommand && tempCommand.getPermissions().length != 0) {
-                        fullPermissions.push({
-                            id: cmd.id,
-                            permissions: [...permissions]
-                        });
-                    }
-                });
-
-                await tempGuild.commands.permissions.set({
-                    fullPermissions
-                });
+                await registerPermissions(guild.id, client);
             });
         } catch (err: any) {
             throw new Error(err.message);
