+++ src/botfiles/utils/registry.ts
@@ -1,10 +1,8 @@
 import path from 'path';
 import { promises as fs } from 'fs';
 import DiscordClient from '../client/client';
-import { SlashCommandBuilder } from '@discordjs/builders';
-const { REST } = require('@discordjs/rest');
-const { Routes } = require('discord-api-types/v9');
-const { BOT_TOKEN, CLIENT_ID, GUILD_ID } = process.env;
+import { PermissionResolvable } from 'discord.js';
+import { RolePermission } from '../../@types/server';
 
 const commandsArray = [];
 export async function registerCommands(client: DiscordClient, dir: string = '') {
@@ -37,5 +35,50 @@ export async function registerEvents(client: DiscordClient, dir: string = '') {
             client.events.set(event.getName(), event);
             client.on(event.getName(), event.run.bind(event, client));
         }
+    }
+}
+
+export async function registerPermissions(guildId: string, client: DiscordClient) {
+    try {
+        const tempGuild = await client.guilds.resolve(guildId);
+        if (!tempGuild) {
+            return;
+        }
+        const permissions: RolePermission[] = [];
+        const fullPermissions: any[] = [];
+        client.commands.forEach(async (command) => {
+            if (command.getPermissions().length === 0) return;
+            const roles = tempGuild.roles.cache.filter((r) =>
+                r.permissions.has(command.getPermissions() as PermissionResolvable)
+            );
+
+            roles.forEach((r) => {
+                permissions.push({
+                    id: r.id,
+                    type: 'ROLE',
+                    permission: true
+                });
+            });
+        });
+        const commands = await client.application?.commands.fetch({
+            guildId: tempGuild.id
+        });
+
+        commands?.forEach(async (cmd) => {
+            const tempCommand = client.commands.get(cmd.name);
+
+            if (tempCommand && tempCommand.getPermissions().length != 0) {
+                fullPermissions.push({
+                    id: cmd.id,
+                    permissions: [...permissions]
+                });
+            }
+        });
+
+        await tempGuild.commands.permissions.set({
+            fullPermissions
+        });
+    } catch (error) {
+        console.log(error);
     }
 }
